added a utils file for homeFeed
