+++ ChatSX/store/reducers/homefeedReducer.ts
@@ -1,8 +1,5 @@
-import {
-  removeDuplicates,
-  sortChatrooms,
-} from '../../components/TaggingView/utils';
 import Styles from '../../constants/Styles';
+import {removeDuplicates, sortChatrooms} from '../../utils/homeFeedUtils';
 import {
   ACCEPT_INVITE_SUCCESS,
   GET_DMFEED_CHAT_SUCCESS,
